Setup test metadata for CircleCI.
This sets up junit XML output for CircleCI so we can better track
progress of which tests pass/fail.

As part of this, switch to mocha-multi for the reporter so we can get
console output as well as the XML file at the same time. This led to a
switch from grunt-mocha-test to grunt-mocha-cli, since mocha wasn't
exiting properly when using mocha-multi in the context of grunt.
grunt-mocha-cli uses a spawning approach for running grunt which seems
more reliable and means we don't have to track down whatever async
process is preventing exit (similar to
https://github.com/glenjamin/mocha-multi/issues/3).

Also remove some unused code coverage stuff that's defunct on this lua
branch.