Fix error handling in lua branch when the error data is invalid.
This fixes some of the newer tests we added surrounding this type of
invalid data.

The behavior of null error data is changing to return the default error
data instead of an internal server error (which is actually what the
test describes, but not what it does). This is easier here since the
null values get stripped in lua and don't perform an override. And I
think this behavior is actually somewhat preferable.