Flip the URL for the mongodb tests to ensure the db config is loaded.
Changing the order of the URL path helps ensure we don't fall back to
any default /info APIs but instead ensure that the API we're hitting is
actually loaded from the db config.

Also change the order of when we reset the mongodb data back in the test
suite so it happens after we've restored the mongodb connection.
Hopefully this will help with intermittent test errors when the reset
happens too soon.