Try to fix intermittent config load race conditions causing test errors.
Also add some better error logging when mongodb or elasticearch queries
fail.

The basic change is to defer setting the active APIs until after dyups
has setup the backends. This will hopefully help with 502 bad gateway
errors we sometimes see in the test environment due to the frequent
config changes and the barrage of /state API requests coming in before
the backend is setup via dyups.

It still appears like a race condition is possible since we don't really
know when dyups has made the backend active across all worker processes,
so we've also increased the delay after dyups adds things to hopefully
help.

This also fixes a few spots in the test suite where we weren't properly
waiting for config to load before running the tests.

This also fixes some testing failures after the mongodb replicaset
changes. These were caused when the nodejs mongodb client would try to
reset the database config, but fail because it wasn't connected to the
correct primary. This seems like more of an issue with the mongoose
adapter not properly picking up the new primary gracefully, but since
it's just the test connection, we'll retry these queries against mongo.