Fix order of host matching.
This fixes several things involving which hosts and apis get matched in
which order. This still could use a lot of cleanup, but this new
approach should be much more consistent. It also fixes unknown redirect
handling for unknown hosts. Hosts are matched in order from longest to
shortest (most specific to least specific, this is consistent with
nginx's host matching). APIs are matched in the sorted API order without
regard to hostname matching order (since we have a distinct matching
order concept among the APIs).