Fix logging of special characters and invalid utf-8 in lua branch.
The logging tests involving invalid utf-8 started failing when we
upgraded the lua utf8 library in 285c3dcea73be8624ddc2d8c67c0dd1b65d4675f

However, I think the real issue was that our previous approach wasn't
really ever working properly and was only passing our tests due to a bug
in the utf8 library's escape method. So instead, we'll take a slightly
different approach and normalize things in the URL so any non-ascii
characters in the URL get URL encoded. While perhaps not the most ideal
solution (it would be nice if we could leave valid UTF-8 strings as-is),
but it seems like the simplest solution to cleanly solve this in all
cases.

This also fixes a previous problem with logging requests with invalid
non-UTF-8 characters in the path portion of the URL.

Also add more tests surrounding all this special character handling.