Remove extraneous text authorization header test under lua branch.
It turns out the behavior of this type of test is rather ambiguous under
nginx. Whether the username portion gets parsed depends on how the
base64 string gets encoded, so it may work for some strings, but not
others. Since this behavior is actually invalid (it was just being
tested for overall behavior when we were parsing our own authorization
headers in nodejs), let's just remove this test.