Fix some finicky tests. Default to no global nginx rate limits.
The changes in the test suite should hopefully help improve the
reliability of some of the tests that have been finicky and periodically
fail depending on the environment.

The file descriptor test should be improved by ensuring we wait for the
reload to complete before checking for open file descriptors. This means
that only one set of nginx worker processes will be running, rather than
possibly two sets, which I think is what periodically led to most of the
issues.

The nginx global rate limit tests should be improved by explicitly
lowering their limits while testing to rates that should be easier and
more reliable to test against (since we should easily be able to achieve
10r/s even on slow/busy hardware). As part of this, we're also going to
ditch the header-based approach to disabling these global rate limits in
the test environment. Now that we have an easy way to reload all config
during the test suite, this makes it a lot easier to change these global
limits on the fly which removes the need for the funky header solution
we were previously reliant on.

We're also going to disable the global nginx rate limits by default. We
may get rid of them entirely (since they can be confusing with the API
Umbrella rate limits), but I think for now, just disabling them by
default is the better choice so the option is still there for anybody
that wants them, but you have to better understand the implications of
enabling them.