Fix race condition in config loading during hyper-frequent nginx reloads
This fixes the race condition uncovered in
627bf24daefeb036c11f57bfa996ae9adda71591. While this wasn't likely to be
encountered, this should properly address the various possible race
conditions that could happen when nginx is reloaded super-frequently and
multiple nginx process groups are spinning up and shutting down at the
same time.

This also attempts to bring some better organization to all of the
config loading code so things are better separated and it's easier to
follow. But I think this could still use some further cleanup.