Fix some race conditions related to config reloading during tests.
During the test suite if a test involved both file and database config
changes and the database config got processes before the file config,
this could lead to failing tests. This was only really an issue in the
test environment where we might have frequent reloads when we're
changing the file config and database config at the same time. This
cropped up in some of the DNS integration tests where the file config
changes with custom TTL settings weren't getting read in before the
database config was, which meant the DNS queries might have happened
with the default DNS config settings (rather than the intended ones that
were part of the associated file config).

We can solve this by waiting for the file config changes to kick in
before proceeding with the database config changes. This does lead to
slightly longer wait times, but since the file config changes don't
happen super-frequently in the test suites (usually just a few times or
less per file), this hopefully won't have a big impact.