Add headers to allow for caching API backends that set HTTP basic auth.
This is part of https://github.com/18F/api.data.gov/issues/281

The bulk of this will be implemented (with tests) over in the router
project, since that's where the caching stuff lives. But this allows us
to distinguish between authorization headers we add due to api backend
config versus externally sent authorization headers.