Fix handling of long hostnames.
It was possible that nginx would bomb if an API backend was added with a
long hostname. The exact length that would trigger this error varied,
since nginx's default lengths actually varied depending on the CPU type.
But this should fix things so that we should properly support hostnames
up to 110 characters long regardless of CPU defaults.

Fixes https://github.com/NREL/api-umbrella/issues/168