Changes to removed hard-coded path defaults. Also remove unused config.
These changes switch all the config that involved paths to happen later
during the configuration processing. This allows us to base all of our
default paths on the API_UMBRELLA_INSTALL_ROOT environment variable,
which should make it easier to test or run things against a non-default
installation location.

This also removes a variety of old configuration items and a few test
things that are no longer relevant in the lua branch.