Fix routing for web APIs. Add install for static site.
The internal API Umbrella APIs coming from the web-app weren't being
routed to in the lua branch. That should be fixed, and tests added for
that routing.

This also adds installation and setup for the basic static site example
(for the home page and api key signup). It's a little funky, since we
have to create an api key for it to function.

Also add database seeding on startup for the initial admin accounts.