Fix empty basic auth header being appended to some requests.
If the basic auth settings int the config was an empty string, then an
empty basic auth header was being set on the requests. Ideally we
wouldn't have empty strings in our config settings, but our production
database does seem to have this for some api backends, so let's work
around this.

Address a similar (but less impactful) issue when the query strings to
append was an empty string (this was resulting in an extra trailing "&"
at the end of the URL).