Don't lowercase all overridden response headers.
It turns out our lowercasing wasn't necessary, since the
ngx.resp.get_headers() returns a special table that works regardless of
case (it just appears lowercase when inspecting it). Lowercasing the
header names shouldn't really matter since header names are case
insensitive (and nodejs always lowercases them, so it's a bit difficult
to test), this just fixes a poorly written case-sensitive check I had in
place, and in looking at this again, I realized we don't need to
actually be doing this.