Interval Lock utility
Use openresty's shared memory (with expiration) functionality to implement
time-based locks. Function calls can be wrapped so that they are only executed
once in a given period of time. Rather than requiring functions to track their
own locked status, the onus is placed on the caller, allowing the wrapped
function to be called explicitly if needed or placed behind multiple locks.