Fix missing seed database data for the admin app to run.
The api key for the web admin to use for ajax calls wasn't being seeded.
The admin permissions collection also wasn't being seeded.

Shift this seeding into the lua startup scripts, since we're no longer
going to call the rake seed command in the web app (we're still keeping
the seeds there for running standalone tests in the web app, but we
should eventually clean this up).