Fix memory growth due to geoip2 module. Add IPv6 support for geoip.
We're seeing unbounded memory growth due to the geoip2 module lookups.
Revert to the built-in nginx geoip module that uses the legacy data
format (but maxmind still provides these and maintains the legacy data
formats, so I don't think this should be a big deal).

Also switch to using the IPv6 dataset download, so that we can geocode
ipv6 addresses (tests added).

Also fix an issue with api-umbrella-geoip-auto-updater not being able to
run the "api-umbrella" reload command on dataset updates to the
less-privileged user it was running as (we may want to eventually
revisit this so that the api-umbrella user can trigger reloads, but
that's a bit more involved at this point).

One note about geoip2's memory usage, is that I had actually picked it
due to seemingly better memory characteristics in some original testing:
https://github.com/NREL/api-umbrella/commit/50258e09bcaf5347d65e26970f95882acdbfcc32
What I saw previously with the geoip2 module was that it started the
processes without much memory use, but then slowly grew as more IP
addresses were used. So some growth was to be expected, but unlike the
initial benchmarks where I'm pretty sure I saw things level off, we seem
to be seeing unbounded growth on production.

In contrast, the builtin geoip module loads the entire city database
into memory at start, so the initial memory use is higher, but then
we're not seeing any growth. Furthermore, after a bit more
investigating, I think the builtin geoip module has better memory use
overall, despite the initially higher usage. Each worker process reports
higher memory usage, which looks even worse when there's lots of worker
processes running, but it turns out this memory is shared across all the
worker processes (due to forking+copy-on-write):
https://forum.nginx.org/read.php?29,159998,160063#msg-160063 Whereas,
I'm not totally sure, but I believe the geoip2 module does not behave
this way, and each worker process does contain it's own copy of the
geoip dataset, so the memory use can actually be much higher with lots
of worker processes.