Fix nginx socket leak test due to geoip changes.
Or at least I'm pretty sure it's the geoip module addition that's
started to cause this test to consistently fail. However, the socket
growth isn't of any real concern, I think it's just that something else
in geoip also holds open a socket to /dev/urandom, so it causes some
small variations in the number of sockets open.