Reduce number of nginx workers used during tests. Improve leak test.
Only use 2 nginx workers in the test environment so there's less
contention and we get better performance when running tests on a
smaller VM with less CPU cores.

Improve the nginx descriptor leak test so that we properly wait for the
reload to fully spin up the new worker group. Instead of just looking
for the number of workers to match the expected number, we now ensure
the old worker processes are actually shut down.