Fix the lua proxy not picking up api user edits.
If an API key was edited to add roles, change rate limits, etc, these
changes were not being picked up by the proxy if the key had already
been seen and was in the local cache.

We had tests for this scenario that were passing, but they weren't quite
valid, since there was a disconnect between how the tests did updates
and how the real application performed updates. The basic issue is that
we now rely on a special "ts" (mongodb timestamp) field on the user
records to detect changes, but the Rails app wasn't setting this field
on updates.

The rails app has been updated to properly set this "ts" field on
updates and we've added integration tests to ensure we pickup changes
that are made via the web APIs. We've added similar tests to ensure API
backend publishing also works when done via the API, rather than just in
the tests (though this was already working).