Try to fix web tests now that we need MongoDB 2.6+ for $currentDate.
CircleCI's default system wide MongoDB is 2.4, so the new tests that
rely on $currentDate don't work.

I'd still like to figure out a better way to run these web tests within
the context of our "make test", but in the meantime, we'll see if we can
get the CircleCI tests running.