Fix some timeout race conditions that could sometimes cause test errors.
Because the timeouts in each proxy was the same, it was possible for
different pieces of the proxy to timeout which could cause some
unexpected behavior in the timeout tests (notably, in some occasional
cases, trafficserver could end up triggering a request retry).

By staggering the timouts at each proxy layer, this should hopefully
help eliminate the possibility of these race conditions. The impact of
this shouldn't be huge in the real world (requests will still timeout),
but should hopefully fix some sporadic tests failures.