Lower proxy timeouts for text environment.
Now that the proxy timeouts are more easily configurable, let's lower
them in the test environment so we don't always have to take 60 seconds
to do integration tests related to timeouts. Now in the test
environment, the default proxy timeout is 10 seconds. That should still
be long enough to be quite easily testable, but no need to run the
timeout tests for a full minute.