Add way to install lua vendored dependencies during deploys.
Without this we can't do live deploys from git if they involve updates
to the lua dependencies.

As a result, this fixes the luarocks binaries and files getting
installed into the system with incorrect paths. There were some
hardcoded references to the build staging directory that need to be
manually replaced when we install our version of the files into the
destination directory.

This also improves the installation of the tar.gz based lua dependencies
so that if the versions change the Makefile detects that and will
install a new version (previously it wasn't aware of the version so if
any version was already installed nothing was done).