Make sure the tmp directory always exists to fix chmod on this dir.
We recently started chmoding this dir on startup. This worked the first
time in the CI environment (since we create this directory as part of
our make process), but it's failing on subsequent tries, since this
directory won't exist if the make process isn't triggered again. So
force this directory to always be part of a git clone.