Add flags to cli reload command to only reload certain components.
We also now need these in our tests to only reload the router
components, rather than also the web components. Some of the tests
started to fail when we added the ability to reload the web components
in, since the web components can take much longer to spin up than the
tests expect. So we'll bring back the reload "--web" and "--router"
flags we used to have and use those in the tests to only reload the
router components.