Fix trafficserver preventing circular requests. Improve Via headers.
TrafficServer was preventing API Umbrella from making circular requests
to itself as an API backend. This is probably a very niche case, but we
did have an old legacy setup using this type of configuration. Fix this
by not passing the Via headers on the requests.

Not passing the Via headers on the requests also potentially helps with
the decent possibility that API backends won't return gzipped content to
us if this header is present. The bandwidth and speed gains of having
API backends gzip the data could be substantial.

Also improve the Via header on the response. Remove the TrafficServer
version number and replace the hostname with a generic alias.

Also improve the regex efficiency of the Via header matching for
generating the X-Cache header.