Updating packaging scripts to build packages across other linux distros.
Now we can build packages for CentOS 6, CentOS 7, Ubuntu 12.04, Ubuntu
14.04, Debian 7, and Debian 8.

The one functional change is that we're no longer building OpenResty
with dtrace support. I ran into issues getting this to work easily
across distros. It does allow a couple of other debugging tools to be
used, but I think there are actually systemtap equivalents, so I'm not
sure if it's really necessary.