Add the concept of a build revision for openresty.
This is so that if we make changes to the ./configure or make process
for openresty, we can manually trigger a rebuild by bumping the build
revision number. This fixes scenarios like when we enabled dtrace in the
openresty configure, but since the version number of openresty didn't
change, an already cached and built system wouldn't rebuild openresty.

We should make similar changes to other packages if we ever modify their
build process without changing versions.