Fix nokogiri security issue by getting rid of nokogiri prod dependency.
Recent libxml security issue affecting nokogiri:
https://bugzilla.gnome.org/show_bug.cgi?id=746048

omniauth-cas was the only thing using nokogiri in production (it's still
used in the test environment for capybara). So let's switch to a fork
that removes the nokogiri dependency altogether. The fact that nokogiri
compiles its own version of libxml by default mean it's somewhat
non-obvious that libxml vulnerabilities necessitate a nokogiri upgrade
(and upgrading the system version of libxml does not help). So let's try
this as an alternative.