Re-include nokogiri in production dependencies for premailer-rails.
It turns out premailer-rails needs nokogiri or hpricot, it's just not an
explicit gem dependency. So we need to add this back in now that
omniauth-cas isn't including it.

And unfortunately, I think this is a tricky one to test, which is why
our test suite didn't complain, since capybara happens to also pull in
nokogiri.