Fix the health check never getting to green with the new interval locks.
The elasticsearch setup was never completing successfully since the HTTP
calls it was making on the initial pass were done directly in the
init_worker_by_lua context. Since HTTP APIs are disabled in this
context, things were failing. To fix, we'll wrap all of our interval
lock calls in an immediately executing timer to shift things into the
timer context. This enables the HTTP APIs, and also makes the context
for the function consistent between the first call and subsequently
schedules calls. This is how we previously scheduled the first wrapper
call to check() instead our spawn() functions.