Fix the distributed rate limit pusher not firing as often as needed.
This was causing some test failures, since each worker process was not
syncing its local data as frequently as expected. The rate limit pusher
is a bit of an odd case amongst our background workers, since it's the
one place where we don't actually want a mutex or locking of any
fashion. So we've reverted its timer setup to the previous setup that
doesn't use the new interval_lock utility, but adding a comment about
why this is done differently.