Fix CSV downloads of request logs if timestamps are milliseconds.
After the lua upgrade we now end up logging the request_at field as a
number, rather than an ISO8601 string. Internally, elasticsearch is
still treating them as dates, but for some reason the JSON output seems
to match the input (which has shifted from strings to numbers). This was
breaking the CSV downloads, since the time parsing was failing. Now time
parsing should be able to cope with dates being returned as either
strings or integers.