Another attempt at fixing admin permissions API.
This is really strange, because the previous fix didn't actually do the
trick, and now I can't reproduce the behavior I'm seeing on production
locally for development or test purposes. So something else appears to
be at play, but respond_with is weird. It would be nice to get to the
bottom of this, but let's give this a shot.