Another attempt to really fix the admin permission and roles APIs.
It turns out the user roles API was similarly broken in production. But
we have decent integration coverage of that API, so now I'm really
perplexed as to what's happening on production and how the output could
be so fundamentally different.

In any case, these seem to be the only two places we were using
"respond_with" for an array of results, which seems to be the issue. So
let's get rid of respond_with altogether here.