Fix editing custom rate limits duplicating the custom rate limit.
In the API User admin, if you attempted to change the value of a custom
rate limit and update the user, the old limit would remain in place,
while the new rate limit would become a duplicate entry.

This was affecting custom rate limits on users, but not API backends.

This was being caused by the user show JSON endpoint accidentally
returning the rate limit data with an "_id" field instead of the
expected "id" field. This caused the PUT to update the data to omit any
IDs, which is why the admin always thought the updates were new records.