Fix rate limit deletes on users by changing timestamping implementation
It turns out the recent issues with custom rate limits on users
(duplicate adds and broken removals), stem from the overriden save()
method that we added to the user model as part of the Lua architecture
changes. The basic goal was to maintain a server-side timestamp of when
user records changed, but the upsert approach turned out to cause other
bugs due to complexities with stuff Mongoid normally does behind the
scenes. To fix this, we'll go back to the default save() implementation
and maintain this timestamp with a second, separate database call. This
isn't as optimized, but it lets us leave the normal save method alone
for Mongoid to do its normal voodoo.