Allow for nginx's server_names_hash_bucket_size option to be configured
And clarify change in behavior in the tests. Now this option should only
need to be changed if a host is explicitly defined in the config that
exceeds the system's default limit. Hosts added via API Backends do not
require this setting to be changed (since we no longer define actual
server{} blocks for each hostname in the nginx config).