Various cleanup for ORC file output.
Tests seem to indicate ORC has the best performance for us. It also
offers the best compatibility with the older version of Hive we're using
for HDP 2.2/Kylin compatibility (allowing us to support things like
timestamps, which aren't supported on the older version of Parquet).

This cleans up our dependencies, removes all the file outputs except
ORC, and tunes a few of the output fields to be specific to ORC files
(using SMALLINTs, etc).