Updating heka for logging to HDFS. Work in progress...
The basic idea is that we'll still use heka to log to possibly different
sources (elasticsearch, postgres, or hdfs). For HDFS (for use with Hive
and Kylin), we'll perform logging from Heka to Flume, which has better
integration with writing directly to HDFS.

This updates the various logged field names for better consistency with
the SQL design of the table.