Some initial experiments with logging to Kafka via rsyslog.
This is for eventually logging to Hive/Kyling via a custom process
pulling off of Kafka. The switch to rsyslog is due to several bugs
encountered with Haka's Kafka output. But rsyslog also has built-in
ElasticSearch and Postgres output, so it may actually be a better option
if we want to eventually log to postgres as an option.