Rework some of the log transfer and storage details.
This is hopefully the beginning of the initial strategy we'll take for
getting these logs into Hive and Kylin on a live-ish basis.

This also tweaks some of the storage to clarify that the minutes field
is also time-zone aware (since some timezones are fractional hours), and
also introduces the request_at_tz_week as a field we'll group by (since
we need that for weekly graphs).