Fix cmake compatibility with older versions of cmake.
The URL_HASH functionality wasn't added until cmake 2.8.10. In order to
build on the default version of cmake for ubuntu 12.04 (cmake 2.8.7),
we'll reorder things. This effectively disables the URL hash checking
when building under this version, but since Ubuntu 12.04 isn't supported
for too much longer, this seems like the simplest workaround for now.