Fix rails tests not finding new config settings. Redo config integration
The rails tests weren't picking up the new default `analytics.timezone`
option. But now that the web app is in the same repo as the top-level
default.yml config, we can solve this in a better way. So rather than
replicating parts of the default config inside the web-app, leverage the
real default.yml for defaults settings inside the rails YAML setup.
Still prefer the API_UMBRELLA_RUNTIME_CONFIG environment variable for
setting the config location, but tweak some of the behavior for more
logical behavior (we won't read in the default runtime_config.yml file,
even if it exists, in the test environments, so tests don't pickup
development settings).