Update packaging process and add packaging for hadoop-analytics.
We're reverting back to using FPM for creating binary packages for a
couple reasons:

1. I ran into bugs with CPack handling file paths containing "%"
   characters (they weren't being escaped properly in the RPM spec
   file).
2. FPM does a bit more normalization across platforms for us out of the
   box, so we don't have to do things like various RPM spec hacking in
   CPack to prevent the rpmbuild from doing things like repacking jar
   files and other odd default behavior.

So while we lose a bit of the appeal of using CMake & CPack for
everything, I think the conversion to CMake still helped cleanup our
build process quite a bit, and CMake's concept of components still makes
it easier to build multiple packages.

So we're back to FPM, but the packaging script has been turned into a
CMake script, and we're now generating the optional hadoop-analytics
package.