Fix possibility of negative TTLs in distributed rate limit puller.
Previously, the TTL in mongo was inadvertently set to 1 minute. However,
for bigger rate limit buckets (eg, with a duration of 24 hours and an
accuracy of 10 minutes), this 1 minute TTL would cause negative TTL
calculations causing errors when syncing the data locally. We now bas
the mongo TTLs on the bucket duration.