Fix website backends listing being visible for limited admins.
The listing of website backends didn't properly account for admin
permissions in some cases. But the admins couldn't edit or display the
individual website backends, so the issue was limited only to the
listing page.

See: https://github.com/18F/api.data.gov/issues/261

This also adds a "none" scope for mongoid, which a couple of our
policies were already erroneously using (which was resulting in errors
being thrown in certain unexpected admin permission circumstances).