Fix tests due to geo IP result changes.
There were some small shifts in some of the geoip results in the latest
data file. Fix things so we're not quite as sensitive to these small
changes in the future (we'll allow results within 0.02 decimal places,
since that's we don't actually care that much about the detailed
precision).

Also fix one of the foreign tests with a special character, since that
IP no longer geocodes to the city. We'll use Bogotá instead.