Fix dependency on non-existent fork of mail gem.
The fork of the mail gem we were depending on went away from github.
This was because the pull request actually wasn't valid at patching the
security issue. However, we had already addressed this security issue in
a different way, so we didn't actually need to depend on this fork
anyway.