Reorganize testing docs, and remove defunct docs.
This consolidates our docs on testing to the documentation website, but
adds links to them from the README. It also clarifies the status of
testing with docker (marked as experimental since it doesn't allow for
testing the full app).

This also removed some defunct documentation left over from when the
web component was a separate repo and played a slightly different role.