Use JSON instead of TSV for Hadoop log transfer.
We had issues with columns being mis-aligned in the TSV due to the
partition columns not being accounted for. Since it now looks like Kylin
1.5 requires JSON Kafka messages for it's streaming support, let's go
ahead and switch to JSON message (we're not currently using the Kylin
streaming support, but JSON's much easier to reason with and deal with
new fields in any case).