Fix drilldown prefix matching under Elasticsearch 2. Add tests.
Building upon dd938d6a25e109225c5b119a8644685ef55bd1a1

For the terms aggregations, we still do need to escape the regex,
however the "^" prefix can no longer be used. It looks like this
"includes" argument was already being anchored automatically, so while
the "^" prefix was unnecessary in Elasticsearch 1, it worked. In
Elasticsearch 2, it now appears as though it must *not* be used (but the
query is still anchored as we want).

This also adds more tests to the drilldown API to test these changes, as
well as more general tests for the drilldown API that we were missing.