Switch analytics logging of "request_at" to be end time (not start time)
Previously, the "request_at" field in the analytics was the timestamp of
when the request began. This changes the behavior so "request_at" is now
the timestamp of when the response finished.

Using the start time caused potential issues with the new hadoop
analytics processing for long-running requests. Because we partition the
live incoming data by the request_at timestamp, it meant that if a
request took 3 minutes to complete, we would suddenly be inserting data
into the partition from 3 minutes ago. This complicates data
processing, since we have to deal with out-of-order insertion timestamps
(which vary depending on how long the requests take).

By switching to use the end time, data processing is much simpler, since
the timestamps should always be incrementing.

Using the end time also better aligns our timestamps with the timestamps
nginx logs in the access.log file (which can be useful for debugging).