Rename SQL columns of logs from "request_at_*" to "timestamp_*".
In light of https://github.com/NREL/api-umbrella/pull/251, the field
we're storing is now the end time, rather than beginning time of the
request. To make this column name more generic, rename the fields to
"timestamp_*" in the new SQL storage. Also better clarify the main
"timestamp_utc" column to clarify it should always contain the UTC time
(versus the "timestamp_tz_*" columns).

Elasticsearch will continue to store it as "request_at" for backwards
compatibility.