Add Ubuntu 16.04 packaging. Fix and improve some packaging things.
The packaging makefile is now more flexible with specific targets for
building a single distro's packages. The download packages script is
also improved to only download the necessary packages.

Also fix an erroneous --rpm-dist flag when building .deb packages.

We also need a newer version of librdkafka for Ubuntu 16.04
compatibility, but the newer v0.9.1 seems to compile fine against
rsyslog.