Remove config import/export tool from the admin.
The export tool was previously open to a security issue for limited
admins (see 9a88f7e68e4be378a5fe1c9f3c5018d3851b4634). However, that
also uncovered that the import part of this tool is mostly broken, and
has been for quite a while (maybe 1-2 years?).

Since this tool is mostly non-functional in its current state, let's go
ahead and remove it entirely. Since nobody's apparently using this (or
mentioning it's brokenness), it doesn't seem worth fixing the current
implementation. Plus, if anyone actually needs this functionality, we
should really redo it in an API-ified way with an Ember interface to
align with the rest of the admin tool (also preferably with admin-scope
aware permissions).