Fix geocoding test for city-state that has recently changed in geoip.
We previously had a test for Singapore to test for city-state situations
where geoip would return a city and country, but no region. However,
Singapore has now started returning a region, so let's switch to testing
against Monaco which is still returning the way we want this test to
run.