Implement dirty form tracking/warning.
This functionality wasn't actually working in the old app since making
the transition to the initial Ember app (since the previous jquery
dirtyform plugin was only being called once on document ready). But now
we'll use an ember approach based on tracking actual changes to the
model to determine whether to warn about navigating away.