Upgrade web-app to Rails 4.2. (Work in progress)
This tackles the bulk of the Rails 4.2 upgrade, and cleans up various
things (particularly since we've removed the Ember app and UI from the
Rails app). The primary outstanding issue is that some of the
multi-level deep associations on the Api class aren't saving properly
due to a bug in the newer versions of Mongoid:
https://jira.mongodb.org/browse/MONGOID-4222 So we'll still need to
address that somehow, but I wanted to go ahead and commit, since this
had already ballooned into a messy change-everything type of commit.

Other things changed in addition to the general Rails 4.2 stuff:

- Also upgrade Ruby to 2.3.
- Also upgrade MongoDB to 3.2, since the old Ruby moped adapter was the
  blocker for moving to a newer version.
- Remove all the defunct assets from the web-app (since all that has
  been shifted to the admin-ui component).
- Remove defunct MyUSA gem and authentication pieces.
- Some cleanup and style tweaks based on the newer version of Rubocop.
- Improve logging setup so that we log to STDOUT when running in
  development mode (so the puma log files can be found in the standard
  /var/log/api-umbrella/web-puma location)
- Improve omniauth setup by only requiring the gems that are actually
  enabled.