Add additional tests to clarify users API behavior involving arrays.
This mainly tests the behavior of how the array fields on API user
settings can get unset in different situations (by either setting it to
null or an empty array). (I think this is isn't working in the Rails 4
upgrade branch, so I wanted to add additional tests around the current
behavior).