Cleanup and simplify assets for login page.
Since the web-app's only assets are down to the simple login page, this
removes the need for jQuery and Bootstrap on these pages.

This also cleans up the styling of the login page a bit after the
bootstrap 3 upgrade.

Also fixes a long-standing bug where Persona would redirect to
non-existent "#{failure_path}" after login failures (rather than a real
url). Just in time for us to eventually remove Persona...