Fix static site in development environment mode.
There was nothing setup to serve the example static site when in
development mode. I think at some point (or maybe we never implemented
this), the middleman preview server would run on this port. But since
that doesn't appear to be the case, let's just point to the built
version for now.