Disable logging the "request_query" object in ElasticSearch by default.
We don't actually do anything with this data currently, and it's caused
various issues (most recently 524d2da2fd797e5c289433148be93a9a901da2bd).

Rolling out the most recent type fixes in 524d2d are also complicated by
the fact that you'd need to reindex data in the current index that's not
normalized to use strings. Without reindexing data, the newly normalized
data may still fail to index in cases where the data was first seen as
an array/object type (since the new flattened strings are incompatible
with those types).

Since we don't have any user-facing uses of this field (it's mainly
still around due to our usage in the test suite), go ahead and disable
its gathering altogether. But we'll still offer a flag to enable this if
anyone needs it (but they'll have to deal with reindexing data types).