Rename test classes to be consistent with filenames.
Enable Rubocop's "ExpectMatchingDefinition" option to ensure that our
test class names match the filenames. This helps catch situations where
we copy and past test class definitions, but forget to update the class
name. This could potentially lead to oddities if the same class is
reopened multiple times for different types of tests.